Refactor: redirectCount 범위가 더 넓은 정수 타입  Long 으로 변경

+++ shortenurlservice/src/main/java/kr/co/shortenurlservice/domain/ShortenUrl.java
@@ -3,7 +3,7 @@ package kr.co.shortenurlservice.domain;
 public class ShortenUrl {
     private String originalUrl;
     private String shortenUrlKey;
-    private int redirectCount;
+    private Long redirectCount;
 
     public ShortenUrl() {}
 
@@ -14,7 +14,7 @@ public class ShortenUrl {
     public ShortenUrl(String originalUrl, String shortenUrlKey) {
         this.originalUrl = originalUrl;
         this.shortenUrlKey = shortenUrlKey;
-        this.redirectCount = 0;
+        this.redirectCount = 0L;
     }
 
     public String getOriginalUrl() {
@@ -25,7 +25,7 @@ public class ShortenUrl {
         return shortenUrlKey;
     }
 
-    public int getRedirectCount() {
+    public Long getRedirectCount() {
         return redirectCount;
     }
